refactor: rename folders
