+++ src/pages/favorites/favorites.tsx
