+++ src/pages/favorites/index.ts
